Update macros.cfg - update LOAD_FILAMENT and UNLOAD_FILAMENT macros
Currently, the LOAD_FILAMENT and UNLOAD_FILAMENT macros in Klipper heat the hotend to 220°C, which is suitable for PLA but may not be optimal for other materials like PETG. 

When using PETG, the extruder has to work much harder than necessary.

However, BambuLab machines employ a more advanced logic. They heat the hotend to 250°C for any type of material, regardless of what is currently being loaded. 

This is because it's crucial to melt not only the new filament but also any material that may have been left in the nozzle from the previous print (could be ASA, PLA, ABS, PC...).

After loading or unloading the filament, the hotend starts cooling down to 180°C to prevent oozing, which is another smart feature.

I have personally implemented this macro on my RatRig machines, and it has worked flawlessly. I believe it can greatly benefit the RatRig community as well, and I'm excited to share it with you.